routes for login/logout and register were made.
