+++ app.js
@@ -1,14 +1,14 @@
 const express = require('express');
 const path = require('path');
-const register = require('./api/registerRouter');
 const app = express();
 const mongoose = require('mongoose');
 const AuthRoute = require('./routes/auth')
 const session = require('express-session');
-const User = require('./models/user')
-const bcrypt = require('Bcryptjs')
-const jwt = require('jsonwebtoken')
 const MongoDBStore = require('connect-mongodb-session')(session);
+const login = require('./routes/login');
+const register = require('./routes/register');
+const logout = require('./routes/logout');
+
 mongoose.Promise = global.Promise
 
 const mongoDBUri = 'mongodb+srv://PosterApp:A1s2f4g5%2A@poster-data.noemv.mongodb.net/Data'
@@ -27,7 +27,6 @@ const store = new MongoDBStore({
     uri: mongoDBUri,
     collection: 'database'
 });
-app.use(express.urlencoded({extended: true}));
 
 app.use(session({
     secret: 'slob that bob top',
@@ -43,75 +42,19 @@ const isAuth = (req, res, next) => {
         res.redirect('/')
     }
 }
-
-app.get('/homepage',isAuth , (req, res) => {
-    res.sendFile(path.resolve(__dirname,'public/homepage/homepage.html'))
-})
-
 app.get('/', (req, res) => {
     res.sendFile(path.resolve(__dirname,'public/login/login.html'))
-})
-
-app.post('/', async (req, res) => {
-    const { email, password} = req.body;
-
-    const user = await User.findOne({email});
-
-    if (!user) {
-        return  res.redirect('/');
-    }
-
-    const isMatch = await bcrypt.compare(password, user.password)
-
-    if (!isMatch) {
-        return  res.redirect('/');
-    }
-
-    req.session.isAuth = true
-    res.redirect('/homepage');
 })
 
 app.get('/register', (req, res) => {
     res.sendFile(path.resolve(__dirname,'public/register/register.html'))
 })
 
-app.post('/register', async (req, res) => {
-    const {
-        username,
-        password,
-        email,
-        lastName,
-        firstName
-        } = req.body
-
-    let user = await User.findOne({email})
-
-    if (user) {
-        return res.redirect('/register');
-    }
- 
-    const hashedPass = await bcrypt.hash(password, 12);
-
-    user = new User({
-        username,
-        password: hashedPass,
-        email,
-        lastName,
-        firstName
-    })
-    await user.save();
-    res.redirect('/')
-});
-
-
-app.post('/logout', (req, res) => {
-    req.session.destroy((err) => {
-        if (err) throw err;
-        res.redirect('/');
-    })
+app.get('/homepage',isAuth , (req, res) => {
+    res.sendFile(path.resolve(__dirname,'public/homepage/homepage.html'))
 })
 
-const port = process.env.PORT || 5000;
+const port = process.env.PORT || 8000;
 app.listen(port, () => console.log(`Listening on port ${port}...`));
 
 
@@ -120,3 +63,6 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use('/api',AuthRoute)
 app.use('/static', express.static(path.join(__dirname, 'public')))
+app.use(login)
+app.use(register)
+app.use(logout)
