--- controller/usersController.js
@@ -1,16 +1,0 @@
-const User = require('../models/user');
-
-
-const index = (req, res, next) => {
-    User.find()
-    .then(response => {
-        res.json({
-            response
-        })
-    })
-    .catch(error => {
-        res.json({
-            message: 'an error Occured'
-        })
-    })
-}
